+++ src/network/postgres_protocol_handler.cpp
@@ -724,7 +724,7 @@ size_t PostgresProtocolHandler::ReadParamValue(
       auto peloton_type = PostgresValueTypeToPelotonValueType(
           static_cast<PostgresValueType>(param_types[param_idx]));
       bind_parameters[param_idx] =
-          std::make_pair(peloton_type, std::string(""));
+          std::make_pair(static_cast<int>(peloton_type), std::string(""));
       param_values[param_idx] =
           type::ValueFactory::GetNullValueByType(peloton_type);
     } else {
@@ -734,7 +734,7 @@ size_t PostgresProtocolHandler::ReadParamValue(
         // TEXT mode
         std::string param_str = std::string(std::begin(param), std::end(param));
         bind_parameters[param_idx] =
-            std::make_pair(type::TypeId::VARCHAR, param_str);
+            std::make_pair(static_cast<int>(type::TypeId::VARCHAR), param_str);
         if ((unsigned int)param_idx >= param_types.size() ||
             PostgresValueTypeToPelotonValueType(
                 (PostgresValueType)param_types[param_idx]) ==
@@ -757,7 +757,7 @@ size_t PostgresProtocolHandler::ReadParamValue(
               int_val = (int_val << 8) | param[i];
             }
             bind_parameters[param_idx] =
-                std::make_pair(type::TypeId::INTEGER, std::to_string(int_val));
+                std::make_pair(static_cast<int>(type::TypeId::INTEGER), std::to_string(int_val));
             param_values[param_idx] =
                 type::ValueFactory::GetIntegerValue(int_val).Copy();
             break;
@@ -768,7 +768,7 @@ size_t PostgresProtocolHandler::ReadParamValue(
               int_val = (int_val << 8) | param[i];
             }
             bind_parameters[param_idx] =
-                std::make_pair(type::TypeId::BIGINT, std::to_string(int_val));
+                std::make_pair(static_cast<int>(type::TypeId::BIGINT), std::to_string(int_val));
             param_values[param_idx] =
                 type::ValueFactory::GetBigIntValue(int_val).Copy();
             break;
@@ -781,14 +781,14 @@ size_t PostgresProtocolHandler::ReadParamValue(
             }
             PL_MEMCPY(&float_val, &buf, sizeof(double));
             bind_parameters[param_idx] =
-                std::make_pair(type::TypeId::DECIMAL, std::to_string(float_val));
+                std::make_pair(static_cast<int>(type::TypeId::DECIMAL), std::to_string(float_val));
             param_values[param_idx] =
                 type::ValueFactory::GetDecimalValue(float_val).Copy();
             break;
           }
           case PostgresValueType::VARBINARY: {
             bind_parameters[param_idx] = std::make_pair(
-                type::TypeId::VARBINARY,
+                    static_cast<int>(type::TypeId::VARBINARY),
                 std::string(reinterpret_cast<char *>(&param[0]), param_len));
             param_values[param_idx] = type::ValueFactory::GetVarbinaryValue(
                 &param[0], param_len, true);
